set `created_on` when object is instantiated
Previously, we were relying on the ORM to set the `created_on`
timestamp if one was not provided via `captured_on`. This happens
during object save, but we needed it earlier than that.

This commit sets `created_on` early in the `__init__` so it can be
referenced later.
