+++ run_unit_tests.sh
@@ -9,7 +9,7 @@ psql 'postgresql://insights:insights@localhost:5432' -c 'create database testdb;
 
 BASELINE_DB_NAME=testdb FLASK_APP=historical_system_profiles.app:get_flask_app_with_migration flask db upgrade
 
-BASELINE_DB_NAME=testdb prometheus_multiproc_dir=$TEMPDIR nosetests -sx --with-coverage --cover-package historical_system_profiles  --cover-min-percentage 59 --cover-erase && rm -rf $TEMPDIR
+BASELINE_DB_NAME=testdb prometheus_multiproc_dir=$TEMPDIR nosetests -sx --with-coverage --cover-package historical_system_profiles  --cover-min-percentage 75 --cover-erase && rm -rf $TEMPDIR
 
 result=$?
